+++ src/HTTPie/Abstractions/IHttpParser.cs
@@ -9,7 +9,8 @@ public interface IHttpParser
 {
     string? Environment { get; set; }
 
-    Task<HttpRequestMessage> ParseScriptAsync(string script, CancellationToken cancellationToken = default);
+    IAsyncEnumerable<HttpRequestMessageWrapper> ParseScriptAsync(string script,
+        CancellationToken cancellationToken = default);
 
     IAsyncEnumerable<HttpRequestMessageWrapper> ParseFileAsync(string filePath,
         CancellationToken cancellationToken = default);